Title: 2차원 최대 증가 수열, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 231220/2차원 최대 증가 수열/longest-increasing-sequence-2d.cpp
@@ -9,8 +9,8 @@ int main() {
   // freopen("output.txt", "w", stdout);
   int n, m;
   cin >> n >> m;
-  vector<vector<int>> grid(n, vector<int>(m, 0));
-  vector<vector<int>> dp(n, vector<int>(m, 0));
+  vector<vector<short>> grid(n, vector<short>(m, 0));
+  vector<vector<short>> dp(n, vector<short>(m, 0));
 
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < m; j++) {
